Update Alerts.xml
Used the English and German file to compare, and filled in the missing information. Cleaned up the language so it was not so awkward Google-translate-y, corrected typos etc. Not entirely sure on object names (e.g. "joy", "medical bed", "tailor station" etc) agreed upon, but will keep working on the whole thing and if needed go back and get everything up to standard.